pyproject.toml
[tool.poetry]
name = "ots-crawl-3d"
version = "0.1.0"
description = ""
authors = ["Tran Nhan <nhantt.gis@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
pandas = "^2.2.3"
playwright = "^1.49.0"
requests = "^2.32.3"
pydantic-settings = "^2.6.1"
fake-bpy-module = "^20241129"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
rich = "^13.9.4"
black = "^24.10.0"
pylint = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '\.git|\.hg|\.mypy_cache|\.tox|\.venv|build|dist'
skip-source-first-line = false
skip-string-normalization = false
skip-magic-trailing-comma = false
